workaround for bug in Docker
If you pipe stderr to anything, the first time you attempt to write to it,
Docker will immediately exit your process (making it return code 0, to
boot, making it very hard to diagnose). This is fixed by... Not piping to
stderr. Luckily, the Boost build does not produce anything on stderr anyways!

ip:debloat