Stop relying on <elf.h> being installed.
This "solution" is not ideal, but requiring users to find their own
elf.h is an even worse approach. I copied the values we expect out of
the elf.h that we downloaded as part of our CI pipeline and pasted them
into the file where they're needed.

Properly find capstone when it has been built and installed, rather than found via the binary release.