getFunctionName() no longer disambiguates function names
getFunctionName() was only used for the purpose of checking
whether a function should be skipped when using the `dynamic`
policy. If multiple copies of a function that should be skipped
are present in a GTIRB file, the symbols for startup functions that should
be skipped are ambiguous and get renamed, meaning the policy
is effectively ignored. The disambiguation thus serves no purpose,
and only makes the function not work.