avoidRegisterNameConflicts now a part of IntelSyntax
Most GCC assemblers have no problem with symbols sharing
names with registers or other asm keywords, and compile
correctly. Intel is the major exception, and so
now only IntelSyntax::formatSymbolName adds '_renamed' so
as to avoid conflicts.

Because of this quirk in Intel syntax, it might be advisable
to have ATT syntax be the default for the pretty-printer instead.