No longer escaping single quotes
clang-as explicitly crashes when single quotes are
escaped, and gas silently ignores the backslash before any escape
sequence (such as a single quote) that it doesn't  recognize.
Assembler::escapeByte() therefore should not escape single quotes.

Since overriding this behavior was the only thing distinguishing
subclasses of Assembler, it is possible we should remove the entire
component.