Fix fakebin test bugs
A couple of bugs caused test failures in the Windows fakebin tests for
me locally:

1. Because PeBinaryPrinter searched for a binary that was not present
in fakebin, it expected to fall back to the default values. However, I
have uasm installed, so it found uasm and executed that instead of
executing fakebin binaries. I updated PeBinaryPrinter to search for
"ml.exe" instead of "cl" to resolve this; this should be reliable on
Windows, correctly match the fakebin regardless of OS, and actually
reflects the binary that we will use.

2. run_binary_pprinter_mock_out captures stdout by default (intended to
suppress output) when running under unittest, but not under pytest. The
test test_windows_pe_resource_data reads stdout, so failed under pytest
because stdout was None. I enabled `check_output` in this test to ensure
stdout was always captured.