Revert the change in printVersionScript
Previously, UND symbols were included in the version-script for
resolving the linking error about no symbol version section.

However, the problem occurred only with --dummy-so because the generated
dummy library was problematic.

To resolve the problem, printVersionScript now takes an additional arg
so that the call printVersionScript in generating dummySo, the version
script can have just version defs without symbols.