Add ALIGN property to SEGMENT containing blocks that require alignment
This change fixes MasmPrettyPrinter so that ALIGN property can be added
to SEGMENT when it contains any data blocks that require alignment.

However, `_DATA` SEGMENT is an exception because it is one of the
predefined SEGMENTs whose properties cannot be changed.
Masm fails when a data block within `_DATA` has an ALIGN directive
bigger than the SEGMENT alignemnt (16).
To get around the problem, it adjusts such ALIGN directives to match to
the SEGMENT alignment value.

However, if blocks within the data section are rearranged after transforms,
alignment issues may arise.
For instance, a block that requires 32-byte alignment could end up
being aligned to only 16-bytes.